requirements.txt
matplotlib==3.7.1
numpy==1.23.5
scipy==1.10.1
biopython==1.81
ete3==3.1.3
pandas==1.5.3
setuptools==67.7.2
PyYAML==6.0.1
